+++ src/apis/instance.js
@@ -1,7 +1,6 @@
 import axios from 'axios';
 
 const backendUrl = process.env.REACT_APP_BACKEND_URL;
-console.log(backendUrl);
 
 const instance = axios.create({
   baseURL: backendUrl,
